LPS-123628 Populate regions during company lifecycle listener instead of country model listener since this will only be executed once when a new company is registered
