+++ modules/apps/address/address-impl/src/main/java/com/liferay/address/internal/portal/instance/lifecycle/PortalInstanceLifecycleListenerImpl.java
@@ -17,15 +17,19 @@ package com.liferay.address.internal.portal.instance.lifecycle;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.instance.lifecycle.BasePortalInstanceLifecycleListener;
 import com.liferay.portal.instance.lifecycle.PortalInstanceLifecycleListener;
+import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONFactory;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.Company;
+import com.liferay.portal.kernel.model.Country;
+import com.liferay.portal.kernel.model.Region;
 import com.liferay.portal.kernel.model.Release;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.service.CountryLocalService;
+import com.liferay.portal.kernel.service.RegionLocalService;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.util.StringUtil;
 
@@ -85,13 +89,15 @@ public class PortalInstanceLifecycleListenerImpl
 
 				serviceContext.setUserId(defaultUser.getUserId());
 
-				_countryLocalService.addCountry(
+				Country country = _countryLocalService.addCountry(
 					countryJSONObject.getString("a2"),
 					countryJSONObject.getString("a3"), true, true,
 					countryJSONObject.getString("idd"), name,
 					countryJSONObject.getString("number"), 0, true, false,
 					countryJSONObject.getBoolean("zipRequired"),
 					serviceContext);
+
+				_processCountryRegions(country);
 			}
 			catch (Exception exception) {
 				if (_log.isDebugEnabled()) {
@@ -112,6 +118,67 @@ public class PortalInstanceLifecycleListenerImpl
 		return _jsonFactory.createJSONArray(regionsJSON);
 	}
 
+	private void _processCountryRegions(Country country) {
+		String a2 = country.getA2();
+
+		try {
+			String path =
+				"com/liferay/address/dependencies/regions/" + a2 + ".json";
+
+			if (getClassLoader().getResource(path) == null) {
+				return;
+			}
+
+			JSONArray regionsJSONArray = _getJSONArray(path);
+
+			if (_log.isDebugEnabled()) {
+				_log.debug("Regions found for country " + a2);
+			}
+
+			for (int i = 0; i < regionsJSONArray.length(); i++) {
+				try {
+					JSONObject regionJSONObject =
+						regionsJSONArray.getJSONObject(i);
+
+					ServiceContext serviceContext = new ServiceContext();
+
+					serviceContext.setCompanyId(country.getCompanyId());
+					serviceContext.setUserId(country.getUserId());
+
+					Region region = _regionLocalService.addRegion(
+						country.getCountryId(),
+						regionJSONObject.getBoolean("active"),
+						regionJSONObject.getString("name"), 0,
+						regionJSONObject.getString("regionCode"),
+						serviceContext);
+
+					JSONObject regionLocalizationsJSONObject =
+						regionJSONObject.getJSONObject("localizations");
+
+					if (regionLocalizationsJSONObject != null) {
+						for (String key :
+								regionLocalizationsJSONObject.keySet()) {
+
+							_regionLocalService.updateRegionLocalization(
+								region, key,
+								regionLocalizationsJSONObject.getString(key));
+						}
+					}
+				}
+				catch (PortalException portalException) {
+					if (_log.isWarnEnabled()) {
+						_log.warn(portalException, portalException);
+					}
+				}
+			}
+		}
+		catch (Exception exception) {
+			if (_log.isDebugEnabled()) {
+				_log.debug("No regions found for country " + a2, exception);
+			}
+		}
+	}
+
 	private static final Log _log = LogFactoryUtil.getLog(
 		PortalInstanceLifecycleListenerImpl.class);
 
@@ -120,6 +187,9 @@ public class PortalInstanceLifecycleListenerImpl
 
 	@Reference
 	private JSONFactory _jsonFactory;
+
+	@Reference
+	private RegionLocalService _regionLocalService;
 
 	@Reference(
 		target = "(&(release.bundle.symbolic.name=portal)(release.schema.version>=9.2.0))"
